+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingStageStatus/proposalVotingStageStatus.test.tsx
@@ -67,6 +67,15 @@ describe('<ProposalVotingStageStatus /> component', () => {
         expect(screen.getByTestId(IconType.CLOSE)).toBeInTheDocument();
     });
 
+    it('correctly renders the expired state', () => {
+        const status = ProposalVotingStatus.EXPIRED;
+        render(createTestComponent({ status }));
+        expect(screen.getByText('Proposal')).toBeInTheDocument();
+        expect(screen.getByText('has')).toBeInTheDocument();
+        expect(screen.getByText('expired')).toBeInTheDocument();
+        expect(screen.getByTestId(IconType.CLOSE)).toBeInTheDocument();
+    });
+
     it('correctly renders the unreached state', () => {
         const status = ProposalVotingStatus.UNREACHED;
         render(createTestComponent({ status }));